refactor(web): extract DashboardRangeContext to break 3 import cycles
The dashboard tab components imported DashboardRangeContext from
../page.tsx while page.tsx imported the tab components, creating
circular dependencies. Extract the context to a dedicated context.ts
file that both sides import from.
