+++ web/src/app/(admin)/dashboard/context.ts
@@ -1,0 +1,6 @@
+// SPDX-FileCopyrightText: 2026 Vishnu Muthiah <vishnu.muthiah04@gmail.com>
+// SPDX-License-Identifier: AGPL-3.0-only
+
+import { createContext } from "react";
+
+export const DashboardRangeContext = createContext<string>("30d");
